Placement unit tests
Added unit tests for placement algorithms. Cleaned up test struture.
